requirements.txt
google-generativeai>=0.2.0
python-dotenv
PyMuPDF
